[#383] Breakdown case settings into components
