+++ frontend/src/pages/cases/Cases.js
@@ -440,6 +440,7 @@ const Cases = () => {
       <CaseSettings
         open={caseSettingModalVisible}
         handleCancel={() => setCaseSettingModalVisible(false)}
+        enableEditCase={true}
       />
     </ContentLayout>
   );
